+++ src/modules/psychologist/logic/verificationOfProfessionals/verificationPsychologist.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { EPsychologistStatus } from '../../enums/verified.enum';
@@ -18,34 +18,6 @@ export class VerificationPsychologistService {
     private readonly psychologistRepository: Repository<Psychologist>,
     private readonly paginationService: PaginationService,
   ) {}
-
-  async getAllVerifiedRequestService(
-    paginationDto: PaginationDto,
-  ): Promise<PaginatedResponse<ResponseProfessionalDto>> {
-    const queryBuilder =
-      this.psychologistRepository.createQueryBuilder('psychologist');
-    queryBuilder.where('psychologist.verified = :status', {
-      status: EPsychologistStatus.PENDING,
-    });
-
-    const paginatedResult = await this.paginationService.paginate(
-      queryBuilder,
-      paginationDto,
-    );
-
-    const transformedItems = plainToInstance(
-      ResponseProfessionalDto,
-      paginatedResult.data,
-      {
-        excludeExtraneousValues: true,
-      },
-    );
-
-    return {
-      ...paginatedResult,
-      data: transformedItems,
-    };
-  }
 
   async getAllVerifiedService(
     paginationDto: PaginationDto,
@@ -72,70 +44,6 @@ export class VerificationPsychologistService {
     return {
       ...paginatedResult,
       data: transformedItems,
-    };
-  }
-
-  async findOne(
-    id: string,
-  ): Promise<{ message: string; data: ResponseProfessionalDto }> {
-    const psychologist = await this.psychologistRepository.findOne({
-      where: { id, is_active: true },
-    });
-
-    if (!psychologist) {
-      throw new NotFoundException('No se encontró el psicólogo');
-    }
-
-    if (psychologist.verified !== EPsychologistStatus.PENDING) {
-      throw new NotFoundException(
-        'El psicólogo ya está verificado o no está pendiente de verificación',
-      );
-    }
-
-    psychologist.verified = EPsychologistStatus.VALIDATED;
-    const savedPsychologist =
-      await this.psychologistRepository.save(psychologist);
-
-    const transformedPsychologist = plainToInstance(
-      ResponseProfessionalDto,
-      savedPsychologist,
-      {
-        excludeExtraneousValues: true,
-      },
-    );
-
-    return {
-      message: 'Psicólogo verificado exitosamente',
-      data: transformedPsychologist,
-    };
-  }
-
-  async rejectPsychologistById(
-    id: string,
-  ): Promise<{ message: string; data: ResponseProfessionalDto }> {
-    const psychologist = await this.psychologistRepository.findOne({
-      where: { id, is_active: true },
-    });
-
-    if (!psychologist) {
-      throw new NotFoundException('No se encontró el psicólogo');
-    }
-
-    psychologist.verified = EPsychologistStatus.REJECTED;
-    const savedPsychologist =
-      await this.psychologistRepository.save(psychologist);
-
-    const transformedPsychologist = plainToInstance(
-      ResponseProfessionalDto,
-      savedPsychologist,
-      {
-        excludeExtraneousValues: true,
-      },
-    );
-
-    return {
-      message: 'Psicólogo rechazado exitosamente',
-      data: transformedPsychologist,
     };
   }
 }
